src/producer/requirements.txt
kafka-python>=2.0.2
prometheus-client>=0.17.0
